Revert "Kill initApp.js in favor of a single app entry point."
