+++ apps/Gruntfile.js
@@ -378,6 +378,7 @@ module.exports = function (grunt) {
     'raceInterstitial':             './src/sites/studio/pages/raceInterstitial.js',
     'termsInterstitial':            './src/sites/studio/pages/termsInterstitial.js',
     'makerlab/setupPage':           './src/sites/studio/pages/setupMakerlab.js',
+    'initApp/initApp':              './src/sites/studio/pages/initApp.js',
     'scriptOverview':               './src/sites/studio/pages/scriptOverview.js'
   };
 
@@ -392,7 +393,7 @@ module.exports = function (grunt) {
     embedVideo: './src/sites/studio/pages/embedVideo.js',
 
     // embedBlocks.js is just React, the babel-polyfill, and a few other dependencies
-    // in a bundle to minimize the amount of stuff we need when loading blocks
+    // in a bundle to minimize the amound of stuff we need when loading blocks
     // in an iframe.
     embedBlocks: './src/sites/studio/pages/embedBlocks.js',
