+++ apps/src/sites/studio/pages/levels-applab-main.js
@@ -1,4 +1,13 @@
-import loadApplab from "./init/loadApplab";
-import loadApp from "@cdo/apps/code-studio/initApp/loadApp";
+import appMain from "@cdo/apps/appMain";
+import Applab from "@cdo/apps/applab/applab";
+window.Applab = Applab;
+if (typeof global !== "undefined") {
+  global.Applab = window.Applab;
+}
+import levels from "@cdo/apps/applab/levels";
+import * as skins from "@cdo/apps/applab/skins";
 
-loadApp(loadApplab);
+window.applabMain = function (options) {
+  options.skinsModule = skins;
+  appMain(window.Applab, levels, options);
+};
