+++ apps/src/sites/studio/pages/levels-bounce-main.js
@@ -1,4 +1,15 @@
-import loadBounce from "./init/loadBounce";
-import loadApp from "@cdo/apps/code-studio/initApp/loadApp";
+import appMain from "@cdo/apps/appMain";
+import Bounce from "@cdo/apps/bounce/bounce";
+window.Bounce = Bounce;
+if (typeof global !== "undefined") {
+  global.Bounce = window.Bounce;
+}
+import blocks from "@cdo/apps/bounce/blocks";
+import levels from "@cdo/apps/bounce/levels";
+import skins from "@cdo/apps/bounce/skins";
 
-loadApp(loadBounce);
+window.bounceMain = function (options) {
+  options.skinsModule = skins;
+  options.blocksModule = blocks;
+  appMain(window.Bounce, levels, options);
+};
