+++ apps/src/sites/studio/pages/levels-calc-main.js
@@ -1,4 +1,12 @@
-import loadCalc from "./init/loadCalc";
-import loadApp from "@cdo/apps/code-studio/initApp/loadApp";
+import appMain from "@cdo/apps/appMain";
+import Calc from '@cdo/apps/calc/calc';
+window.Calc = Calc;
+import blocks from "@cdo/apps/calc/blocks";
+import skins from "@cdo/apps/skins";
+import levels from "@cdo/apps/calc/levels";
 
-loadApp(loadCalc);
+window.calcMain = function (options) {
+  options.skinsModule = skins;
+  options.blocksModule = blocks;
+  appMain(window.Calc, levels, options);
+};
