+++ apps/src/sites/studio/pages/levels-flappy-main.js
@@ -1,4 +1,15 @@
-import loadFlappy from "./init/loadFlappy";
-import loadApp from "@cdo/apps/code-studio/initApp/loadApp";
+import appMain from "@cdo/apps/appMain";
+import Flappy from '@cdo/apps/flappy/flappy';
+window.Flappy = Flappy;
+if (typeof global !== 'undefined') {
+  global.Flappy = window.Flappy;
+}
+import blocks from "@cdo/apps/flappy/blocks";
+import levels from "@cdo/apps/flappy/levels";
+import skins from "@cdo/apps/flappy/skins";
 
-loadApp(loadFlappy);
+window.flappyMain = function (options) {
+  options.skinsModule = skins;
+  options.blocksModule = blocks;
+  appMain(window.Flappy, levels, options);
+};
