+++ apps/src/sites/studio/pages/levels-gamelab-main.js
@@ -1,4 +1,16 @@
-import loadApp from "@cdo/apps/code-studio/initApp/loadApp";
-import loadGamelab from './init/loadGamelab';
+import appMain from "@cdo/apps/appMain";
+import {singleton as studioApp} from '@cdo/apps/StudioApp';
+import GameLab from "@cdo/apps/gamelab/GameLab";
+import skins from "@cdo/apps/gamelab/skins";
+import levels from "@cdo/apps/gamelab/levels";
 
-loadApp(loadGamelab);
+window.gamelabMain = function (options) {
+  options.skinsModule = skins;
+  var gamelab = new GameLab();
+
+  // Bind helper that provides project metadata for gamelab autosave
+  options.getAnimationList = gamelab.getSerializedAnimationList.bind(gamelab);
+
+  gamelab.injectStudioApp(studioApp);
+  appMain(gamelab, levels, options);
+};
