+++ apps/src/sites/studio/pages/levels-jigsaw-main.js
@@ -1,4 +1,15 @@
-import loadJigsaw from "./init/loadJigsaw";
-import loadApp from "@cdo/apps/code-studio/initApp/loadApp";
+import appMain from "@cdo/apps/appMain";
+import Jigsaw from '@cdo/apps/jigsaw/jigsaw';
+window.Jigsaw = Jigsaw;
+if (typeof global !== 'undefined') {
+  global.Jigsaw = window.Jigsaw;
+}
+import blocks from "@cdo/apps/jigsaw/blocks";
+import levels from "@cdo/apps/jigsaw/levels";
+import skins from "@cdo/apps/jigsaw/skins";
 
-loadApp(loadJigsaw);
+window.jigsawMain = function (options) {
+  options.skinsModule = skins;
+  options.blocksModule = blocks;
+  appMain(window.Jigsaw, levels, options);
+};
