+++ apps/src/sites/studio/pages/levels-netsim-main.js
@@ -1,4 +1,14 @@
-import loadNetSim from "./init/loadNetSim";
-import loadApp from "@cdo/apps/code-studio/initApp/loadApp";
+import appMain from "@cdo/apps/appMain";
+import {singleton as studioApp} from '@cdo/apps/StudioApp';
+import NetSim from "@cdo/apps/netsim/netsim";
+import levels from "@cdo/apps/netsim/levels";
+import skins from "@cdo/apps/netsim/skins";
 
-loadApp(loadNetSim);
+window.netsimMain = function (options) {
+  options.skinsModule = skins;
+  options.isEditorless = true;
+
+  var netSim = new NetSim();
+  netSim.injectStudioApp(studioApp);
+  appMain(netSim, levels, options);
+};
