+++ apps/src/sites/studio/pages/levels-weblab-main.js
@@ -1,4 +1,17 @@
-import loadWeblab from "./init/loadWeblab";
-import loadApp from "@cdo/apps/code-studio/initApp/loadApp";
+import appMain from "@cdo/apps/appMain";
+import {singleton as studioApp} from "@cdo/apps/StudioApp";
+import WebLab from "@cdo/apps/weblab/WebLab";
+import skins from "@cdo/apps/skins";
+import levels from '@cdo/apps/weblab/levels';
 
-loadApp(loadWeblab);
+window.weblabMain = function (options) {
+  options.skinsModule = skins;
+  options.isEditorless = true;
+  const weblab = new WebLab();
+
+  // Bind helper that provides project metadata for gamelab autosave
+  // options.getAnimationMetadata = gamelab.getAnimationMetadata.bind(gamelab);
+
+  weblab.injectStudioApp(studioApp);
+  appMain(weblab, levels, options);
+};
