+++ apps/test/integration/util/runLevelTest.js
@@ -99,24 +99,8 @@ StubDialog.prototype.show = function () {
 StubDialog.prototype.hide = function () {
 };
 
-const appLoaders = {
-  applab: require('@cdo/apps/sites/studio/pages/init/loadApplab'),
-  bounce: require('@cdo/apps/sites/studio/pages/init/loadBounce'),
-  calc: require('@cdo/apps/sites/studio/pages/init/loadCalc'),
-  craft: require('@cdo/apps/sites/studio/pages/init/loadCraft'),
-  eval: require('@cdo/apps/sites/studio/pages/init/loadEval'),
-  flappy: require('@cdo/apps/sites/studio/pages/init/loadFlappy'),
-  gamelab: require('@cdo/apps/sites/studio/pages/init/loadGamelab'),
-  jigsaw: require('@cdo/apps/sites/studio/pages/init/loadJigsaw'),
-  maze: require('@cdo/apps/sites/studio/pages/init/loadMaze'),
-  netsim: require('@cdo/apps/sites/studio/pages/init/loadNetSim'),
-  studio: require('@cdo/apps/sites/studio/pages/init/loadStudio'),
-  turtle: require('@cdo/apps/sites/studio/pages/init/loadArtist'),
-  weblab: require('@cdo/apps/sites/studio/pages/init/loadWeblab'),
-};
 function runLevel(app, skinId, level, onAttempt, testData) {
-
-  var loadApp = appLoaders[app];
+  require('@cdo/apps/sites/studio/pages/levels-' + app + '-main');
 
   var studioApp = require('@cdo/apps/StudioApp').singleton;
 
@@ -129,7 +113,8 @@ function runLevel(app, skinId, level, onAttempt, testData) {
     return Boolean(testData.useFirebase);
   };
 
-  loadApp({
+  var main = window[app + 'Main'];
+  main({
     skinId: skinId,
     level: level,
     baseUrl: '/base/build/package/',
