+++ dashboard/app/views/levels/_apps_dependencies.html.haml
@@ -55,11 +55,11 @@
 
 -# Common script dependencies
 %script{src: minifiable_asset_path('js/marked/marked.js')}
+%script{src: minifiable_asset_path('js/initApp/initApp.js')}
 %script{src: minifiable_asset_path('js/common.js')}
 -# common_locale is included in application.html.haml
 %script{src: asset_path("js/#{js_locale}/#{app}_locale.js")}
-%script{src: minifiable_asset_path("js/#{app}.js"),
-        data: {appoptions: app_options.to_json}}
+%script{src: minifiable_asset_path("js/#{app}.js")}
 
 -# Applab external script dependencies
 - if use_applab
