+++ dashboard/app/views/levels/_blockly.html.haml
@@ -22,3 +22,11 @@
 
 - content_for :body_scripts do
   = render_app_dependencies
+
+  -# Small inline script to kick off app initialization once all dependencies
+  -# have been loaded.
+  :javascript
+    //<![CDATA[
+    apps.setupApp(appOptions);
+    apps.load(apps.init);
+    //]]>
