+++ dashboard/app/views/levels/embed_blocks.html.erb
@@ -1,7 +1,6 @@
 <%- # Embedded Blockly block html template. Derived from /blockly/src/templates/readonly.html.ejs, converted to erb %>
 <%-
   options = local_assigns
-  options[:embedded] = true
   app = options.delete :app
 %>
 <!DOCTYPE html>
@@ -14,9 +13,15 @@
   <%= javascript_include_tag "js/#{js_locale}/common_locale" %>
   <script src="<%= minifiable_asset_path('js/essential.js') %>"></script>
   <script src="<%= minifiable_asset_path('js/code-studio-common.js') %>"></script>
-  <%= render partial: 'levels/apps_dependencies', locals: { app: app, use_blockly: true, app_options: options } %>
+  <%= render partial: 'levels/apps_dependencies', locals: { app: app, use_blockly: true } %>
 </head>
 <body class="readonly embed-block">
   <div id="codeWorkspace"></div>
+  <script>
+    var appOptions = <%= raw options.to_json %>;
+    window.onload = function () {
+      <%= app %>Main(appOptions);
+    };
+  </script>
 </body>
 </html>
