+++ dashboard/app/views/peer_reviews/show.html.haml
@@ -20,6 +20,7 @@
     %p Please review the peer submission in the grey box below.
 
   .submitted_readonly{style: 'padding: 8px;'}
+    %script{src: minifiable_asset_path('js/initApp/initApp.js')}
     :javascript
       var appOptions = #{app_options.to_json};
     = render partial: 'levels/free_response', locals: {in_level_group: true}
